add relation for topics table
