--- app/alembic/versions/9afe7105662d_add_module_topic_category.py
@@ -1,46 +1,0 @@
-"""add module topic_category
-
-Revision ID: 9afe7105662d
-Revises: ec76d464675a
-Create Date: 2024-06-03 23:11:16.099485
-
-"""
-from alembic import op
-import sqlalchemy as sa
-import sqlmodel.sql.sqltypes
-
-
-# revision identifiers, used by Alembic.
-revision = '9afe7105662d'
-down_revision = 'ec76d464675a'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('topic_category',
-    sa.Column('id', sa.Integer(), nullable=False, autoincrement=True),
-    sa.Column('category_name', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-
-    op.bulk_insert(
-        sa.table(
-            'topic_category',
-            sa.column('id', sa.Integer),
-            sa.column('category_name', sa.String),
-        ),
-        [
-            {'id': 1, 'category_name': "SOSHUM"},
-            {'id': 2, 'category_name': "SAINTEK"},
-            {'id': 3, 'category_name': "CAT 3"}
-        ]
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('topic_category')
-    # ### end Alembic commands ###
