+++ app/alembic/versions/f30f21e2cbc6_add_topic_category_module_with_relation.py
@@ -1,0 +1,62 @@
+"""add_topic_category_module with relation
+
+Revision ID: f30f21e2cbc6
+Revises: ec76d464675a
+Create Date: 2024-06-04 18:23:55.887712
+
+"""
+from alembic import op
+import sqlalchemy as sa
+import sqlmodel.sql.sqltypes
+
+
+# revision identifiers, used by Alembic.
+revision = 'f30f21e2cbc6'
+down_revision = 'ec76d464675a'
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('topic_category',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('category_name', sqlmodel.sql.sqltypes.AutoString(), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+
+    op.bulk_insert(
+        sa.table(
+            'topic_category',
+            sa.column('id', sa.Integer),
+            sa.column('category_name', sa.String),
+        ),
+        [
+            {'id': 1, 'user_id': 1, 'category_name': 'SOSHUM'},
+            {'id': 2, 'user_id': 2, 'category_name': 'SAINTEK'},
+            {'id': 3, 'user_id': 3, 'category_name': 'CAT 3'},
+        ]
+    )
+    op.create_index(op.f('ix_topics_topic_name'), 'topics', ['topic_name'], unique=False)
+    op.create_foreign_key(None, 'topics', 'topic_category', ['topic_category_id'], ['id'])
+    op.alter_column('user_topic_rating', 'topic_id',
+               existing_type=sa.INTEGER(),
+               nullable=True)
+    op.create_index(op.f('ix_user_topic_rating_user_id'), 'user_topic_rating', ['user_id'], unique=False)
+    op.create_foreign_key(None, 'user_topic_rating', 'topics', ['topic_id'], ['id'])
+    op.create_foreign_key(None, 'user_topic_rating', 'users', ['user_id'], ['id'])
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_constraint(None, 'user_topic_rating', type_='foreignkey')
+    op.drop_constraint(None, 'user_topic_rating', type_='foreignkey')
+    op.drop_index(op.f('ix_user_topic_rating_user_id'), table_name='user_topic_rating')
+    op.alter_column('user_topic_rating', 'topic_id',
+               existing_type=sa.INTEGER(),
+               nullable=False)
+    op.drop_constraint(None, 'topics', type_='foreignkey')
+    op.drop_index(op.f('ix_topics_topic_name'), table_name='topics')
+    op.drop_table('topic_category')
+    # ### end Alembic commands ###
